utils/crawler/requirements.txt
requests
BeautifulSoup4
selenium
python-dotenv